functionality for RPN ex01 is done, just need to check if it is valid or not...

+++ cpp_projs/CPP09/ex01/RPN.hpp
@@ -6,7 +6,7 @@
 #include <stack>
 class RPN{
     private:
-        std::stack<int> queue;
+        std::stack<long long> queue;
     public:
         RPN();
         ~RPN();
